FEATURE: Spinner / loading bar progress and error handling improvements
Vcaller now captures the output of certain applications rather than
flood the user's screen with messages, condensing its progress into
visually appealing spinners or loading bars, depending on whether or not
a process's progress can be measured. For now, this is primarily applied
to post-alignment processing steps. In addition, error handling in
Vcaller is now more robust, capable of dealing with more exceptions.

Additionally, the addition of Read Groups is now a mandatory step.